- Add script "fit_yang_yang.py" along with "yang_yang_1dbg.py", the output depends on "processed_data.h5" and generates "fit_outputs.h5"
"yang_yang_1dbg.py" has a new method that returns g1D,
a.k.a. the regularized coupling constant, present in the Lieb-Liniger Hamiltonian.
This is used later on to change units of density and chemical potential.

"fit_yang_yang.py" has two modes, local and global, and new additions to the
fitting methods include the "break_LDA" flag in the model function for the potential
that is responsible for applying YY thermodynamics within the trapped gas region as
well as the binning of the data as it is loaded before the optimization begins to
increase the speed.

- Add script "make_figures.py" which depends on "fit_outputs.h5" and
"processed_data.h5" as well as "Fig1_Schematic.jpg" (also added) to put
together all the figures that are currently considered for the draft.

"make_figures.py" now contains the temperature calibration done back in june to
convert final_dipole values into 3D-temperatures.